LOL-133 - Cancel button should not delete document

+++ LOL-FE/src/app/components/uploader-home/uploader-home.component.ts
@@ -107,35 +107,40 @@ export class UploaderHomeComponent implements OnInit {
       this.result = dialogResult;
       console.log(this.result)
 
-      docName = { "docName": docName }
-      this.isLoading = true
-      this.manageDocs.deleteDoc(docName)
-        .subscribe({
-          next: (res) => console.log(res),
-          error: (err) => {
-            // DELETE Success
-            if (err.error.text == 'Document deleted!') {
-              // RELOAD upon successful deletion
-              console.log(err.error.text)
-              this.snackbar.open(err.error.text, 'close', {
-                duration: 1000,
-                verticalPosition: "top",
-                panelClass: ["successAlert"]
-              })
-                .afterDismissed().subscribe(() => location.reload())
-            }
-            else {
-              // ADD ERROR snackbar message
-              console.log(err.error.text)
-              this.snackbar.open(err.error.text, 'Close', {
-                duration: 1000,
-                verticalPosition: "top",
-                panelClass: ["errorAlert"]
-              })
-              this.isLoading = false
-            }
-          },
-        });
+      if (this.result == false) {
+        location.reload()
+      } else {
+
+        docName = { "docName": docName }
+        this.isLoading = true
+        this.manageDocs.deleteDoc(docName)
+          .subscribe({
+            next: (res) => console.log(res),
+            error: (err) => {
+              // DELETE Success
+              if (err.error.text == 'Document deleted!') {
+                // RELOAD upon successful deletion
+                console.log(err.error.text)
+                this.snackbar.open(err.error.text, 'close', {
+                  duration: 1000,
+                  verticalPosition: "top",
+                  panelClass: ["successAlert"]
+                })
+                  .afterDismissed().subscribe(() => location.reload())
+              }
+              else {
+                // ADD ERROR snackbar message
+                console.log(err.error.text)
+                this.snackbar.open(err.error.text, 'Close', {
+                  duration: 1000,
+                  verticalPosition: "top",
+                  panelClass: ["errorAlert"]
+                })
+                this.isLoading = false
+              }
+            },
+          });
+        }
 
     });
   }
